@@ -5,7 +5,7 @@ option (COMPILE_WITH_CLANG "COMPILE_WITH_CLANG" OFF)
 
 set_property(GLOBAL PROPERTY USE_FOLDERS ON)
 
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 if (WIN32)
@@ -30,6 +30,9 @@ else (WIN32)
 endif (WIN32)
 
 project (Carve)
+
+add_definitions(-DCARVE_USE_EXACT_PREDICATES)
+add_definitions(-DCARVE_MULTITHREADING)
 
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/bin/)
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/bin/)
